Remove exiting activity list
The list was added since commit 14fbb6e. And migrate
to hierarchical in commit 92cb3fd.

Since an animating container won't be removed from
hierarchy, and will be remove until the callback of
animation finished or handleCompleteDeferredRemoval,
it is unnecessary to use a list to keep the exiting
records.

The traversal order also may not be a problem because
the exiting records are already in background since
hierarchy unifying.

So the list should be no longer needed.

Test: atest DisplayContentTests#testForAllWindows
Bug: 163976519
Change-Id: Ie9c0e7e6c00b619c2d984903beb6c576d9b764ff