Add config to detect e-call call based on dialed sim emerg # list.
We normally use the combined emergency number list from the emergency
number lists of both sims when placing emergency calls, with the provision
that the routing may be unspecified if a dialed number is an emergency
number on one sim but not the other.

This config allows a carrier to mandate that calls are only treated as
an emergency call on their network based on their own emergency number
list.

Test: TBD
Bug: 203947159
Change-Id: Id5cd4123f1bb7bdd85e564fc6997dfbc50652a5f
(cherry picked from commit 18a78149726e9064807f61ff5a75ce7ddf8992bd)