Ignore requests to start an already started system service.
For apex-system-services, we allow them to be explicitly started in
SystemService if there is a dependency on them from the rest of the
framework during dessert development. We also need to preserve their
manifest declaration for any previously launched dessert releases.

To avoid starting the same service multiple times, keep track of which
services have been started in SystemServiceManager and ignore
duplicates.

See go/updatable-system-services#bookmark=id.dnv20ia7siqd for more
details.

Bug: 192880996
Test: presubmit
Change-Id: I0a9e48eac9fd5739e2c35818fa483a8e479c13a1