Reset app freezer timeout for broadcast messages
The current freezer code will unfreeze a frozen app when it receives a
registered broadcast. That grants the app a grace period to process the
message. However, if the app is not already frozen, there isn't such an
extra grace period. It's possible the app might be frozen immediately
after receiving the broadcast.

This patch ensures apps always get that extra grace period to process
the incoming broadcast messages.

Bug: 172001574
Bug: 189609004
Test: cached apps won't be frozen until the grace period expires.

Change-Id: Ie9d8c51b074029dc0b23cf3991f3ab21a556ca17